Adds tests for database version of the web server.

+++ hrs_db.py
@@ -2,18 +2,6 @@ import json
 from pymodm import connect
 from pymodm import MongoModel, fields
 
-# read in credentials, screws with tests.
-try:
-    with open("config.json", 'r') as f:
-        config_info = json.load(f)
-        db_user = config_info["mongo_user"]
-        db_pass = config_info["mongo_pass"]
-except:
-    pass
-
-connect("mongodb://{}:<{}>@ds041337.mlab.com:41337/heart_rate_sentinel".format(
-    db_user, db_pass))
-
 
 class Patient(MongoModel):
     patient_id = fields.CharField(primary_key=True)
@@ -25,7 +13,14 @@ class Patient(MongoModel):
 
 class HRDatabase(object):
     def __init__(self):
-        pass
+        with open("config.json", 'r') as f:
+            config_info = json.load(f)
+            db_user = config_info["mongo_user"]
+            db_pass = config_info["mongo_pass"]
+
+            url = "mongodb://{}:{}@ds041337.mlab.com:41337/heart_rate_sentinel".format(
+                db_user, db_pass)
+            connect(url)
 
     def get_all(self):
         """
@@ -34,7 +29,7 @@ class HRDatabase(object):
             dict: Patients in the database
         """
         ret_json = {}
-        for user in Patient.objects.raw({}):
+        for user in Patient.objects.all():
             ret_json[user.patient_id] = self.convert_to_json(user)
         return ret_json
 
@@ -44,14 +39,13 @@ class HRDatabase(object):
         Args:
             user_info (dict): Dictionary with the user's info.
         """
-        if self.check_user_exists(user_info["patient_id"]):
+        patient = self.get_patient(user_info["patient_id"])
+        if patient:
             raise ValueError("The patient is already in the database.")
 
         p = Patient(patient_id=user_info["patient_id"],
                     attending_email=user_info["attending_email"],
                     user_age=user_info["user_age"],
-                    heart_rates=[],
-                    timestamps=[]
                     )
         p.save()
 
@@ -64,13 +58,12 @@ class HRDatabase(object):
         Returns:
             bool: Whether or not the user was removed.
         """
-        query = {
-            "patient_id": patient_id
-        }
         removed = False
-        for user in Patient.objects.raw(query):
-            removed = True
-            user.delete()
+        # raw won't work for some reason
+        for user in Patient.objects.all():
+            if str(user.patient_id) == patient_id:
+                removed = True
+                user.delete()
         return removed
 
     def get_patient(self, patient_id):
@@ -83,11 +76,10 @@ class HRDatabase(object):
             dict: Information of the patient. Returns None if DNE.
 
         """
-        query = {
-            "patient_id": patient_id
-        }
-        for user in Patient.objects.raw(query):
-            return user
+        # raw won't work for some weird reason...
+        for user in Patient.objects.all():
+            if str(user.patient_id) == patient_id:
+                return user
         return None
 
     def add_hr(self, patient_id, heart_rate, timestamp):
@@ -98,30 +90,17 @@ class HRDatabase(object):
             heart_rate: New heart rate.
             timestamp: New timestamp.
         """
+        user_exists = False
         query = {
             "patient_id": patient_id
         }
-        for user in Patient.objects.raw(query):
-            user.heart_rates.append(heart_rate)
-            user.timestamps.append(timestamp)
-            user.save()
-
-    def check_user_exists(self, patient_id):
-        """
-        Checks if the user exists in the database.
-        Args:
-            patient_id (str): ID
-
-        Returns:
-            bool: Whether or not the user exists in the database.
-
-        """
-        query = {
-            "patient_id": patient_id
-        }
-        if Patient.objects.raw(query):
-            return True
-        return False
+        for user in Patient.objects.all():
+            if str(user.patient_id) == patient_id:
+                user.heart_rates.append(heart_rate)
+                user.timestamps.append(timestamp)
+                user.save()
+                user_exists = True
+        return user_exists
 
     def convert_to_json(self, db_object):
         """
